Allow loading clip arts from user images
This change required another fix: Don't scale bitmaps to 44 pixels wide in convertDrawableToLEDHex. I could not figure out the reason for this, since everything can just handle images that are not 44 pixels wide as long as they are 11 pixels tall, which the new and old code ensures.

The images are converted by scaling them to the largest fitting size, then converting all bright pixels to lit-up and everything dark or transparent to not lit up. This is different from the internal black vs. transparent format, but it appears to be more natural when importing pixel art.